#127 restructured tests and sourcecode
Effectively, moved many-to-one and many-to-many tests into one
"related" file. That way it's easier to have its corresponding code
in postgres in a separate file.

Also the code itself was moved into the sqlite lib.
I realized that some differences in Postgres syntax make it complex
to support both it and sqlite in the same proc.
So now the plan is to add postgres support by copy-pasting the sqlite procs and adjusting them accordingly.