ADD PRAGMA foreign_keys=on to sqlite connections
SQLITE does not enable foreign key constraint checks by default.
They are off by default.
So you can create entry A and entry B depending on A
and then you can delete entry A without an error being thrown.
That is unexpected and behaviour and not consistent with postgres.

Therefore this commit enables the pragma for sqlite connections.
When they are created via `getDb` or added to an sqlite pool

https://github.com/moigagoo/norm/issues/184