Make dbType generic that allows Positive
Nim 1.6.10 used to allow dispatching to a proc
dbType(T: typedesc[int64]) when you
passed in a model with field type "Positive".

This changed in nim 2.
Positive no longer matches int64 this way.
HOWEVER, you also can not define a separate dbType proc for
the "Positive" type.
Because int64 still matches to the Positive type.
So by doing so you would suddenly cause ambiguous
situations where nim can't decide
whether to call dbType for int64 or Positive
when encountering a field of type int64.

Making it a generic this way circumvents that problem.